refactor: scss
move the scss from `styles` to `core/components`
