+++ src/core/ui/components/Card/card.scss
