+++ src/core/ui/components/Loader/loader.scss
