+++ src/core/ui/components/Pagination/pagination.scss
